egg-asn1x: Avoid signed integer overflow
There were a couple of places where signed integer is wrapped around
when overflow, though unlike unsigned integer, it is undefined
behavior:
https://www.gnu.org/software/autoconf/manual/autoconf-2.63/html_node/Integer-Overflow-Basics.html#Integer-Overflow-Basics

Spotted by UndefinedBehaviorSanitizer.

+++ egg/egg-asn1x.c
@@ -701,8 +701,8 @@ atlv_parse_cls_tag (const guchar *at,
                     gulong *tag,
                     gint *off)
 {
-	gint punt, ris, last;
-	gint n_data;
+	guint punt, ris, last;
+	guint n_data;
 	guchar val;
 
 	g_assert (end >= at);
@@ -762,9 +762,9 @@ atlv_parse_length (const guchar *at,
                    const guchar *end,
                    gint *off)
 {
-	gint ans, last;
-	gint k, punt;
-	gint n_data;
+	guint ans, last;
+	guint k, punt;
+	guint n_data;
 
 	g_assert (at != NULL);
 	g_assert (end != NULL);
@@ -809,7 +809,7 @@ atlv_parse_length (const guchar *at,
 		}
 
 		*off = punt;
-		return ans;
+		return (gint) ans;
 	}
 }
